회원 탈퇴 로직 수행 시, 로그인한 회원인지 확인 후 탈퇴 진행하게끔 수정

+++ src/main/java/com/dlrtn/websocket/chat/service/UserService.java
@@ -111,6 +111,10 @@ public class UserService {
 
         User foundUser = userMapper.findByUserId(request.getUserId());
 
+        if (!sessionRepository.exists(sessionId)) {
+            return CommonResponse.failWith("Not Login State");
+        }
+
         if (validateUser(foundUser, request.getPassword())) {
             try {
                 userMapper.delete(user);
